Reverted part of previous commit; rewind states now use 64-bit where appropriate. Fixed some more minor errors found by clang.

+++ src/common/RewindManager.hxx
@@ -73,7 +73,7 @@ class RewindManager
     /**
       Convert the cycles into a unit string.
     */
-    string getUnitString(Int32 cycles);
+    string getUnitString(Int64 cycles);
 
   private:
     // Maximum number of states to save
